Move types used by external services to separate file
